+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This format is based on [Keep a Changelog](http://keepachangelog.com/). This pro
 ## [unreleased]
 ### Added
 - Add attestation to PyPI.
+- Add ``.xyz`` input validation
+- Add tests with fuzzer for input validation.
 ### Changed
 - Change ``typing.py`` to ``types.py`` to prevent name clash.
 ### Fixed